schema change

+++ db/schema.rb
@@ -56,8 +56,8 @@ ActiveRecord::Schema[7.1].define(version: 2024_05_14_182642) do
   end
 
   create_table "leisures", force: :cascade do |t|
-    t.bigint "category_id", null: false
-    t.bigint "venue_id", null: false
+    t.integer "category_id", null: false
+    t.integer "venue_id", null: false
     t.string "picture"
     t.string "link"
     t.string "title"
@@ -78,8 +78,8 @@ ActiveRecord::Schema[7.1].define(version: 2024_05_14_182642) do
   end
 
   create_table "leisures_genres", force: :cascade do |t|
-    t.bigint "leisure_id", null: false
-    t.bigint "genre_id", null: false
+    t.integer "leisure_id", null: false
+    t.integer "genre_id", null: false
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
     t.index ["genre_id"], name: "index_leisures_genres_on_genre_id"
